implimented all changes #26

+++ lib/models/playing_card.dart
@@ -96,7 +96,7 @@ class PlayingCard {
   String typeToString;
   String typeToStringBody;
   Image suitImage;
-  int penaltyVal;
+  double penaltyVal;
   CardColor cardColor;
 
   PlayingCard({
